// autogpt_platform/frontend/src/app/api/__generated__/models/aPIKeyPermission.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */

export type APIKeyPermission =
  (typeof APIKeyPermission)[keyof typeof APIKeyPermission];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const APIKeyPermission = {
  IDENTITY: "IDENTITY",
  EXECUTE_GRAPH: "EXECUTE_GRAPH",
  READ_GRAPH: "READ_GRAPH",
  WRITE_GRAPH: "WRITE_GRAPH",
  EXECUTE_BLOCK: "EXECUTE_BLOCK",
  READ_BLOCK: "READ_BLOCK",
  READ_STORE: "READ_STORE",
  WRITE_LIBRARY: "WRITE_LIBRARY",
  USE_TOOLS: "USE_TOOLS",
  MANAGE_INTEGRATIONS: "MANAGE_INTEGRATIONS",
  READ_INTEGRATIONS: "READ_INTEGRATIONS",
  DELETE_INTEGRATIONS: "DELETE_INTEGRATIONS",
} as const;